rename FiveButtonRating to StarsRatingPanel
