+++ src/modules/feedback/components/rating/starsRatingPanel.css
