+++ src/modules/feedback/i18n/index.js
@@ -1,5 +1,5 @@
 import { provide as feedbackProvider } from './feedback.provider';
-import { provide as fiveButtonRatingProvider } from './fiveButtonRating.provider';
+import { provide as fiveButtonRatingProvider } from './starsRatingPanel.provider';
 import { $injector } from '../../../injection';
 const { TranslationService: translationService } = $injector.inject('TranslationService');
 translationService.register('feedbackProvider', feedbackProvider);
