+++ src/modules/feedback/i18n/starsRatingPanel.provider.js
