+++ src/modules/feedback/index.js
@@ -2,7 +2,7 @@ import './i18n';
 import { MapFeedbackPanel } from './components/mapFeedback/MapFeedbackPanel';
 import { ToggleFeedbackPanel } from './components/toggleFeedback/ToggleFeedbackPanel';
 import { GeneralFeedbackPanel } from './components/generalFeedback/GeneralFeedbackPanel';
-import { FiveButtonRating } from './components/rating/FiveButtonRating';
+import { StarsRatingPanel } from './components/rating/StarsRatingPanel';
 
 if (!window.customElements.get(MapFeedbackPanel.tag)) {
 	window.customElements.define(MapFeedbackPanel.tag, MapFeedbackPanel);
@@ -16,6 +16,6 @@ if (!window.customElements.get(GeneralFeedbackPanel.tag)) {
 	window.customElements.define(GeneralFeedbackPanel.tag, GeneralFeedbackPanel);
 }
 
-if (!window.customElements.get(FiveButtonRating.tag)) {
-	window.customElements.define(FiveButtonRating.tag, FiveButtonRating);
+if (!window.customElements.get(StarsRatingPanel.tag)) {
+	window.customElements.define(StarsRatingPanel.tag, StarsRatingPanel);
 }
